requirements.txt
# jupyter notebook
ipykernel

# langchain init
langchain-openai==0.2.14
langchain==0.3.13
langserve==0.3.0
python-dotenv==1.0.1

# fastapi
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.27.2

# langchain message history
langchain-community==0.3.13

# TextSplitter
lxml==5.3.0
nltk==3.9.1
langchain_experimental==0.3.4

# TextEmbedding
faiss-cpu==1.9.0.post1